🔨 Refactor: 드롭다운, 예약 관련 타입 병합
