+++ src/components/common/modals/ReservationModal.tsx
@@ -4,7 +4,7 @@ import { useConfirmModal } from '@/src/hooks/useConfirmModal';
 import { useReservationModal } from '@/src/hooks/useReservationModal';
 import { formatDateDot } from '@/src/utils/date-format';
 import { MODAL_RESERVATION_STATUS_MAP } from '@/src/constants/reservation-modal';
-import { ModalReservationStatus, ReservationModalProps } from '@/src/types/reservation-modal.types';
+import { ModalReservationStatus } from '@/src/types/reservation.types';
 
 import Dropdown from '../dropdowns/Dropdown';
 import MyReservationsModalCard from '@/src/app/profile/my-reservations/components/MyReservationsModalCard';
@@ -95,4 +95,9 @@ export default function ReservationModal({ activityId, date }: ReservationModalP
       <ConfirmModal />
     </div>
   );
+}
+
+export interface ReservationModalProps {
+  activityId: number;
+  date: Date;
 }
