+++ src/hooks/useReservationModal.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from 'react';
 import { useReservedSchedules } from './useReservedSchedules';
 import { useReservationsByTime } from './useReservationsByTime';
 import { useUpdateReservation } from './useUpdateReservation';
-import { DropdownItemButton } from '../types/dropdown-menu.types';
-import { ModalReservationStatus, ReservedSchedule } from '../types/reservation-modal.types';
+import { DropdownItemButton } from '../types/dropdown.types';
+import { ModalReservationStatus, ReservedSchedule } from '../types/reservation.types';
 
 export function useReservationModal(activityId: number, date: Date) {
   const [reservationStatus, setReservationStatus] = useState<ModalReservationStatus>('pending');
