+++ src/hooks/useReservedSchedules.tsx
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { fetchFromClient } from '../lib/fetch/fetchFromClient';
 import { formatDateHyphen } from '../utils/date-format';
-import { ReservedSchedule } from '../types/reservation-modal.types';
+import { ReservedSchedule } from '../types/reservation.types';
 
 export function useReservedSchedules(activityId: number, date: Date) {
   const formattedDate = formatDateHyphen(date);
