--- src/types/dropdown-item.types.ts
@@ -1,8 +1,0 @@
-export interface DropdownItemProps {
-  label: string;
-  isMobile?: boolean;
-  isDanger?: boolean;
-  disabled?: boolean;
-  onClick: () => void;
-  onClose: () => void;
-}
