--- src/types/dropdown-menu.types.ts
@@ -1,13 +1,0 @@
-export interface DropdownItemButton {
-  label: string;
-  onClick: () => void;
-  isDanger?: boolean;
-}
-
-export interface DropdownMenuProps {
-  items: DropdownItemButton[];
-  bottomSheetTitle: string;
-  isMobile?: boolean;
-  bottomButton?: DropdownItemButton; // 바텀 시트 취소 버튼
-  onClose: () => void;
-}
