--- src/types/dropdown-trigger.types.ts
@@ -1,6 +1,0 @@
-export interface DropdownTriggerProps {
-  label: string;
-  text: string;
-  isOpen: boolean;
-  onClick: () => void;
-}
