+++ src/types/dropdown.types.ts
@@ -1,9 +1,5 @@
-import { DropdownItemButton } from './dropdown-menu.types';
-
-export interface DropdownProps {
-  dropdownItems: DropdownItemButton[];
-  selectedValue: string;
-  bottomSheetTitle: string;
-  trigger?: React.ReactNode;
-  triggerLabel?: string;
+export interface DropdownItemButton {
+  label: string;
+  onClick: () => void;
+  isDanger?: boolean;
 }
