--- src/types/reservation-modal.types.ts
@@ -1,35 +1,0 @@
-export interface ReservationModalProps {
-  activityId: number;
-  date: Date;
-}
-
-export interface ReservedSchedule {
-  scheduleId: number;
-  startTime: string;
-  endTime: string;
-  count: {
-    declined: number;
-    confirmed: number;
-    pending: number;
-  };
-}
-
-export interface Reservation {
-  id: number;
-  nickname: string;
-  userId: number;
-  teamId: string;
-  activityId: number;
-  scheduleId: number;
-  status: string;
-  reviewSubmitted: boolean;
-  totalPrice: number;
-  headCount: number;
-  date: string;
-  startTime: string;
-  endTime: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export type ModalReservationStatus = 'pending' | 'declined' | 'confirmed';
