--- src/types/reservations-status.types.ts
@@ -1,3 +1,0 @@
-import { reservationStatusMap } from '@/src/constants/reservations-status';
-
-export type ReservationStatus = keyof typeof reservationStatusMap;
